instagram/requirements/dev.txt
-r common.txt

django-debug-toolbar
easy-thumbnails